move stats to new file
